moved pags out of test-site folder
