+++ pages/+Page.mdx
@@ -1,15 +1,137 @@
-export { ContentPage as default } from "~/layouts";
-import { macrostratInstance } from "@macrostrat-web/settings";
 import { PageHeader } from "~/components";
-import { LinkCard } from "~/components/cards";
+import { Image, Footer } from "./index";
+import "./main.styl";
+import "./main-page.styl";
+import "@macrostrat/style-system";
 
-<PageHeader>
-  <span className="subtitle">
-    <span className="tertiary">{macrostratInstance}</span>
-  </span>
-</PageHeader>
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous"></link>
+
+<div class="new_container field">
+  <div class="mask">
+    <Image src="main-page/field.jpg" className="start-img back-img"></Image>
+    <div class="main-content row row-eq-height">
+      <div class="side">
+        <div class="center-me">
+            <div class="center">
+              <Image src="main-page/rockd_transparent.png" className="not-huge"></Image>
+              <p>Learn, explore, and document your geologic world</p>
+              <a href="https://itunes.apple.com/us/app/id1153056624"><Image src="main-page/appstore.png" className="download-badge ios"></Image></a>
+              <a href="https://play.google.com/store/apps/details?id=org.macrostrat.rockd"><Image src="main-page/google-play-badge.png" className="download-badge android"></Image></a>
+              <p><a class="explore-link" href="/explore">Explore the data</a></p>
+          </div>
+        </div>
+        <Image src="main-page/main.png" className="main-img"></Image>
+      </div>
+    </div>
+  </div>
+</div>
+
+<div class="new_container map-container">
+  <div class="side">
+    <div class="map-imgs">
+      <Image src="main-page/grand_canyon.jpg" className="map-shot grand_canyon"></Image>
+      <Image src="main-page/new_zealand.jpg" className="map-shot new_zealand"></Image>
+      <Image src="main-page/world.jpg" className="map-shot world"></Image>
+      <Image src="main-page/appalachia.jpg" className="map-shot appalachia"></Image>
+    </div>
+
+    <div class="center map-info">
+      <h1>Instant access to more than 290 <a href="https://macrostrat.org/map/sources">geologic maps</a> globally</h1>
+      <p>Includes direct links to <a href="https://macrostrat.org">Macrostrat</a> and <a href="https://xdd.wisc.edu">xDD</a> (where available)</p>
+    </div>
+  </div>
+</div>
+
+<div class="new_container white-container">
+  <div class="row">
+    <div class="col-sm-6 col-sm-offset-3">
+      <div class="headline-alligner">
+        <div class="center curious">
+          <h1>Curious what you're standing on?</h1>
+          <p>The Dashboard distills key facts about your location into a single, easy-to-read interface</p>
+        </div>
+      </div>
+    </div>
+    <div class="col-xs-12 dashboard-img-container">
+      <Image src="main-page/pilbara-australia.png"></Image>
+      <Image src="main-page/huronian-supergroup.png"></Image>
+      <Image src="main-page/copper-harbor.png"></Image>
+
+      <Image src="main-page/appalachian-foreland.png"></Image>
+      <Image src="main-page/gotthard-pass.png"></Image>
+      <Image src="main-page/ozark-plateau.png"></Image>
+      <Image src="main-page/eastern-australia.png"></Image>
+
+    </div>
+  </div>
+</div>
+
+<div class="new_container red_rock">
+  <Image src="main-page/red_rock.jpg" className="back-img red-rock-img"></Image>
+  <div class="img-container row row-eq-height">
+    <div class="col-sm-6 stats-container">
+      <Image src="main-page/profile.png" className="profile-img"></Image>
+    </div>
+
+    <div class="col-sm-6 center-me">
+      <div class="headline-alligner">
+        <div class="center">
+          <h1>"The best geologist is the one who has seen the most rocks"</h1>
+          <p>Track your progress through the geologic rock record with summary statistics about your checkins</p>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
 
 
-### Development pages
+<div class="record-block">
+  <div class="row">
+    <div class="col-sm-6 col-sm-offset-3">
+      <div class="headline-alligner">
+        <div class="center">
+          <h1>Record your observations</h1>
+          <p>Rockd allows you to easily record your geological observations, using your location to provide spatially informed suggestions for nearby geologic units, time intervals, and fossils</p>
+        </div>
+      </div>
+    </div>
 
-<LinkCard href="/dev" title="Developer apps"></LinkCard>
+    <div class="col-sm-6">
+      <div class="headline-alligner">
+        <div class="center">
+          <Image src="main-page/checkin.png" className="record-block-img"></Image>
+          <p class="record-img-caption"><span>Take pictures of and rate features</span></p>
+        </div>
+      </div>
+    </div>
+
+    <div class="col-sm-6">
+      <div class="headline-alligner">
+        <div class="center">
+          <Image src="main-page/strat-names.png" className="record-block-img"></Image>
+          <p class="record-img-caption"><span>Tag stratigraphic names, even offline</span></p>
+        </div>
+      </div>
+    </div>
+
+    <div class="col-sm-6">
+      <div class="headline-alligner">
+        <div class="center">
+          <Image src="main-page/strike-dip.png" className="record-block-img"></Image>
+          <p class="record-img-caption"><span>Use your phone's compass to record strike and dip</span></p>
+        </div>
+      </div>
+    </div>
+
+    <div class="col-sm-6">
+      <div class="headline-alligner">
+        <div class="center">
+          <Image src="main-page/taxa.png" className="record-block-img"></Image>
+          <p class="record-img-caption"><span>Search for and tag <a href="https://paleobiodb.org">Paleobiology Database</a> taxa.</span></p>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+
+<Footer></Footer>
