+++ pages/colors.scss
