+++ pages/privacy/+Page.mdx
@@ -1,5 +1,5 @@
 import { PageHeader } from "~/components";
-import { Image, Navbar, Footer } from "../index";
+import { Image, Footer } from "../index";
 import "./main.styl";
 import "../main.styl";
 import { MacrostratIcon } from "~/components";
