+++ pages/trip/+Page.client.ts
